platform_python/requirements.txt
about-time==4.2.1
anyio==3.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
fastapi==0.94.1
grapheme==0.6.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
inflate64==0.3.1
IP2Location==8.9.0
jmespath==1.0.1
kafka-python==2.0.2
multivolumefile==0.2.3
nvidia-ml-py==11.525.84
orjson==3.8.7
phone==0.4.3
psutil==5.9.4
py7zr==0.20.4
pyasn1==0.4.8
pybcj==1.0.1
pycryptodomex==3.17
pydantic==1.10.6
pydub==0.25.1
PyMySQL==1.0.2
pypiwin32==223
pyppmd==1.0.0
pyttsx3==2.90
pywin32==305
pyzstd==0.15.4
rarfile==4.0
requests==2.28.2
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.1
texttable==1.6.7
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
